[API] Enable contract tests for statistics (without validation)
