+++ src/Sylius/Bundle/ApiBundle/Resources/config/services/query_handlers.xml
@@ -25,9 +25,9 @@
             <argument type="service" id="sylius.customer_statistics_provider" />
         </service>
 
-<!--        <service id="Sylius\Bundle\ApiBundle\QueryHandler\GetStatisticsHandler">-->
-<!--            <argument type="service" id="Sylius\Component\Core\Statistics\Provider\StatisticsProviderInterface" />-->
-<!--            <argument type="service" id="sylius.repository.channel" />-->
-<!--        </service>-->
+        <service id="Sylius\Bundle\ApiBundle\QueryHandler\GetStatisticsHandler">
+            <argument type="service" id="Sylius\Component\Core\Statistics\Provider\StatisticsProviderInterface" />
+            <argument type="service" id="sylius.repository.channel" />
+        </service>
     </services>
 </container>
